% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fire}
\alias{fire}
\title{Damage from fire, related to distance}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{Distance from fire station in miles}
  \item{DAMAGE}{Fire damage in thousands of dollars}

}
}
\source{
Mendenhall and Schmidt
}
\usage{
fire
}
\description{
A dataset containing the distance form a fire station and the corresponding
damage in dollars.
}
\keyword{datasets}
