% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{My Binomial}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Number of trials}

\item{p}{Probability}
}
\value{

}
\description{
\code{mybin} creates many iterations of a random binomial sample to experimentally see what happens in a binomial distribution.
}
\examples{
mybin(10, 10, 0.6)
}
