% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iterations}

\item{x}{the sample}

\item{fun}{function to create a statistic off of}

\item{alpha}{inverse of confidence interval}

\item{cx}{character size}

\item{...}{other headers for histogram}
}
\value{

}
\description{
\code{myboot2()} creates a bootstrap statistic and associated histogram
}
\examples{
myboot2(x = rnorm(20,10,4))
}
