% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{My Confidence Interval}
\usage{
myci(d, a = 0.05)
}
\arguments{
\item{d}{Data in the form of a list of numbers.}

\item{a}{(1-a)100 \% confidence interval.}
}
\value{

}
\description{
\code{myci} creates a confidence interval for a population mean given a sample.
}
\examples{

myci(c(10,12,13,14,15))

}
