% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{MY Central Limit Theorem}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{the number of samples}

\item{iter}{number of iterations}

\item{a}{the lower bound for the uniform function}

\item{b}{the upper bound for the uniform fucntion}
}
\value{

}
\description{
\code{myclt} creates a statistic from the uniform distribution
}
\examples{
myclt(10,10000,a=0,b=5)
}
