% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My Normal Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The expected value}

\item{sigma}{The standard deviation}

\item{a}{What point we want to evaluate the area less than}
}
\value{

}
\description{
\code{myncurve} creates a curve showing a normal density distribution, then calculates the area under the curve and graphically represents it with a polygon.
}
\examples{
myncurve(3, 1, 4)
}
