% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{My P value}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{The sample}

\item{xmax}{The maximum x value}

\item{n}{The size of the sample}

\item{alpha}{(100-alpha) CI}
}
\value{

}
\description{
\code{mypvalue} creates a p-value plot from a sample.
}
\examples{
mypvalue(qt(1-0.05/2,9), n=10, alpha=0.05)
}
