% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{My Read}
\usage{
myread(x)
}
\arguments{
\item{"Filename.xls"}{}
}
\description{
\code{myread} reads the data from my local excel directory for this class.
}
\details{
The directory is set in the function, but it can be redefined according to what dird needs to be.  The argument searches for a file with an exact filename match, including extension.
}
\examples{
myread("DDT.xls")
}
