% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{My max likelihood}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{The function used}

\item{x}{The data}

\item{param}{Parameter we want to estimate}

\item{...}{other values for the plot}
}
\value{

}
\description{
\code{mymaxlik} creates a maximum likelihood estimator
}
